use startContainer utils
